various fixes

+++ inode.h
@@ -166,7 +166,7 @@ class Inode {
     std::unordered_set<std::string> links_;
 
     /** list of blocks for this inode */
-    std::map<uint32_t, std::shared_ptr<Block>> blocks_;
+    std::map<uint64_t, std::shared_ptr<Block>> blocks_;
